347: moved manager & queryset to phishtray
