+++ participant/managers.py
@@ -1,10 +1,6 @@
 from django.db.models import Manager
 
-from .queryset import OrganizationQuerySet, ParticipantQuerySet
-
-
-class OrganizationManager(Manager.from_queryset(OrganizationQuerySet)):
-    pass
+from .queryset import ParticipantQuerySet
 
 
 class ParticipantManager(Manager.from_queryset(ParticipantQuerySet)):
