+++ participant/models.py
@@ -1,12 +1,8 @@
 from operator import itemgetter
 from statistics import mean
 
-from participant.constants import (
-    ParticipantBehaviour,
-    POSITIVE_ACTIONS,
-    NEGATIVE_ACTIONS,
-)
-from .managers import OrganizationManager, ParticipantManager
+from participant.constants import ParticipantBehaviour, POSITIVE_ACTIONS, NEGATIVE_ACTIONS
+from .managers import ParticipantManager
 
 from django.db import models
 from exercise.models import (
@@ -18,7 +14,7 @@ from exercise.models import (
     ExerciseTask,
 )
 
-from phishtray.base import PhishtrayBaseModel
+from phishtray.base import PhishtrayBaseModel, MultiTenantManager
 
 STARTED_EXPERIMENT = 0
 COMPLETED_EXPERIMENT = 1
@@ -90,9 +86,9 @@ class Participant(PhishtrayBaseModel):
 
         def participant_behaviour(actions):
             pb = ParticipantBehaviour.NEUTRAL
-            if [a for a in actions if a.get("action_type") in POSITIVE_ACTIONS]:
+            if [a for a in actions if a.get('action_type') in POSITIVE_ACTIONS]:
                 pb = ParticipantBehaviour.POSITIVE
-            if [a for a in actions if a.get("action_type") in NEGATIVE_ACTIONS]:
+            if [a for a in actions if a.get('action_type') in NEGATIVE_ACTIONS]:
                 pb = ParticipantBehaviour.NEGATIVE
             return pb
 
@@ -100,7 +96,7 @@ class Participant(PhishtrayBaseModel):
 
         if email_id in self.exercise.phishing_email_ids:
             for action_id, action_details in self.actions.items():
-                if email_id == action_details.get("email_id"):
+                if email_id == action_details.get('email_id'):
                     actions.append(action_details)
 
         return participant_behaviour(actions), actions
@@ -193,4 +189,4 @@ class Organization(models.Model):
     def __str__(self):
         return self.name
 
-    objects = OrganizationManager()
+    objects = MultiTenantManager()
