--- phishtray/managers.py
@@ -1,6 +1,0 @@
-from django.db.models import Manager
-from .queryset import OrganizationQuerySet
-
-
-class OrganizationManager(Manager.from_queryset(OrganizationQuerySet)):
-    pass
